# Conflicts:
#	.vs/ActionGame_ItoDaigo/CopilotIndices/0.2.1627.58842/CodeChunks.db
#	.vs/ActionGame_ItoDaigo/CopilotIndices/0.2.1627.58842/SemanticSymbols.db
#	.vs/ActionGame_ItoDaigo/v17/.suo
#	.vs/ActionGame_ItoDaigo/v17/DocumentLayout.backup.json
#	.vs/ActionGame_ItoDaigo/v17/DocumentLayout.json
#	.vs/ActionGame_ItoDaigo/v17/fileList.bin
#	DX22Base/Component.hxx
#	DX22Base/DX22Base.vcxproj
#	DX22Base/DX22Base.vcxproj.filters
#	DX22Base/GameObjectInst.hxx
#	DX22Base/GameObjectMng.cxx
#	DX22Base/GameObjectMng.hxx
#	DX22Base/Render.cxx
#	DX22Base/Render.hxx
#	DX22Base/SceneBase.hxx
#	DX22Base/SceneMng.hxx
#	DX22Base/Transform.cxx
#	DX22Base/Transform.hxx

+++ DX22Base/TagManager.cpp
@@ -11,7 +11,7 @@ int TagManager::AddTag(const std::string& tagName)
 		if (m_tags[i] == tagName)
 			return i;
 	m_tags.push_back(tagName);
-	return m_tags.size() - 1;
+	return (int)m_tags.size() - 1;
 }
 
 int TagManager::GetTagNo(const std::string& tagName)
